# Speed up `empty_brackets?`