@bbatsov 2023-01-30
I've reviewed the PR myself and given the backwards-compatible signature I see not reason why not to move forward with it. Thanks for tackling this!